+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.2] - 17-10-2025
+
+### Fix
+
+- fix tabs as group button in navbar
+
+### Added
+
+- support tabs which are not text documents
+- support recovery of pinned tabs
+- improve render state
+- add renaming of groups
+- add coloring depending on tab kind
+- add indicator in case tab is not recoverable
+
 ## [1.0.1] - 17-10-2025
 
 ### Fix